Cleanup: move controller_extension into more/
